Ship is now displayed along with background.
