+++ galactic_react/src/components/Stars.jsx
@@ -1,0 +1,12 @@
+import React, { Component } from 'react';
+
+class Stars extends Component {
+  render() {
+    return (
+      <div className="stars">
+      </div>
+    )
+  }
+}
+
+export default Stars;
