+++ galactic_react/src/components/Twinkling.jsx
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+
+class Twinkling extends Component {
+  render() {
+    return (
+      <div className="twinkling">
+        { this.props.children }
+      </div>
+    )
+  }
+}
+
+export default Twinkling;
